+++ lib/src/analyzers/lint_analyzer/lint_analyzer.dart
@@ -5,6 +5,7 @@ import 'package:analyzer/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
+import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/metrics_list/source_lines_of_code/source_lines_of_code_metric.dart';
 import 'package:file/local.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart';
@@ -295,12 +296,12 @@ class LintAnalyzer {
             source,
           );
 
-      final linesOfExecutableCodeVisitor = SourceCodeVisitor(source.lineInfo);
+      final sourceLinesOfCodeVisitor = SourceCodeVisitor(source.lineInfo);
 
-      function.declaration.visitChildren(linesOfExecutableCodeVisitor);
+      function.declaration.visitChildren(sourceLinesOfCodeVisitor);
 
-      final linesOfExecutableCode = MetricValue<int>(
-        metricsId: linesOfExecutableCodeKey,
+      final sourceLinesOfCode = MetricValue<int>(
+        metricsId: SourceLinesOfCodeMetric.metricId,
         documentation: const MetricDocumentation(
           name: '',
           shortName: '',
@@ -308,13 +309,13 @@ class LintAnalyzer {
           measuredType: EntityType.methodEntity,
           examples: [],
         ),
-        value: linesOfExecutableCodeVisitor.linesWithCode.length,
+        value: sourceLinesOfCodeVisitor.linesWithCode.length,
         level: valueLevel(
-          linesOfExecutableCodeVisitor.linesWithCode.length,
+          sourceLinesOfCodeVisitor.linesWithCode.length,
           readThreshold<int>(
             config.metricsConfig,
-            linesOfExecutableCodeKey,
-            linesOfExecutableCodeDefaultWarningLevel,
+            SourceLinesOfCodeMetric.metricId,
+            50,
           ),
         ),
         comment: '',
@@ -357,7 +358,7 @@ class LintAnalyzer {
         (171 -
                 5.2 * log(max(1, halsteadVolume)) -
                 cyclomatic.value * 0.23 -
-                16.2 * log(max(1, linesOfExecutableCode.value))) *
+                16.2 * log(max(1, sourceLinesOfCode.value))) *
             100 /
             171,
       ).toDouble();
@@ -397,8 +398,6 @@ class LintAnalyzer {
             ),
             comment: '',
           ),
-          if (config.metricsConfig.containsKey(linesOfExecutableCodeKey))
-            linesOfExecutableCode,
         ],
       );